Changed zbuffer to float to be more precise

+++ libs/buffer.h
@@ -14,9 +14,9 @@ class TriangularPiece;
 class ZBuffer{
     int XMAX, YMAX;
     public:
-        vector< vector<int> > grid;
+        vector< vector<float> > grid;
         ZBuffer(int, int);
-        void addToBuffer(int, int, int);
+        void addToBuffer(int, int, float);
         void resolveBuffer(TriangularPiece &tri);
         void outputBuffer();
         void clearBuffer();
